+++ ChangeLog.md
@@ -3,5 +3,6 @@
 ## 0.0.2.0
 
 - [#2](https://github.com/parsonsmatt/prairie/pull/2)
-    - Add `tabulateRecordA` to `Record` class. `tabulate` is now a normal function.
+    - Add `tabulateRecordA` to `Record` class. `tabulate` and `allFields` are now normal functions.
+    - Provide a default implementation of `recordFieldLabel` for `Show`able fields.